+++ CHANGELOG.md
@@ -10,8 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 
 ### Changed
 
-* Now uses swords on bamboo and cobweb
-* Now uses shears on cobweb
+* Now uses swords on bamboo and cobweb.
+* Now uses shears on cobweb.
+* Uses item with more attack damage on players or more damage per seconds on mobs.
 
 ### Deprecated